+++ packages/tab/src/index.tsx
@@ -1,0 +1,7 @@
+import './styles.scss';
+
+export * from './Tab';
+export * from './Tabs';
+export * from './TabPanel';
+export * from './TabPanels';
+export * from './TabList';