+++ packages/table/src/TableHead.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 
-interface TableHeadProps extends React.HTMLAttributes<HTMLTableSectionElement> {
+export type TableHeadProps = {
   /** Kolonneoverskrifter */
   children: React.ReactNode;
-}
+};
 
 export const TableHead: React.FC<TableHeadProps> = props => (
   <thead {...props} />