+++ src/gatsby-theme-docz/index.tsx
@@ -7,6 +7,7 @@ import SiteFooter from '~/components/SiteFooter';
 import FrontPageFooter from '~/components/FrontPageFooter';
 import { SettingsProvider } from '~/components/SettingsContext';
 import SettingsPanel from '~/components/SettingsPanel';
+import SEO from '~/gatsby-theme-docz/base/Seo';
 import Props from '~/components/Props';
 import Menu from './UI/Menu';
 import FrontPageMenu from './UI/FrontPageMenu';
@@ -39,6 +40,7 @@ const App: React.FC = ({ children }) => {
           </SkipToContent>
           {isFrontPage ? (
             <>
+              <SEO title="Velkommen" />
               <FrontPageMenu />
               <main id="site-content" className="frontpage-site-content">
                 {children}
@@ -47,6 +49,7 @@ const App: React.FC = ({ children }) => {
             </>
           ) : (
             <>
+              <SEO />
               <Menu />
               <div className="site-content">
                 <main id="site-content">{children}</main>